+++ outlib/Tab.js
@@ -78,7 +78,7 @@ Tab = function (_React$Component) {_inherits(Tab, _React$Component);
 				{ backgroundImage: this.state.favIcon },
 
 
-
+				id: this.props.id,
 				title: this.props.tab.title,
 				onClick: this.click,
 				onMouseDown: this.onMouseDown,